Conflicts:
	Library/Formula/z.rb

+++ Library/Formula/qt.rb
@@ -2,70 +2,72 @@ require 'formula'
 
 class Qt < Formula
   homepage 'http://qt-project.org/'
-  url 'http://releases.qt-project.org/qt4/source/qt-everywhere-opensource-src-4.8.4.tar.gz'
-  sha1 'f5880f11c139d7d8d01ecb8d874535f7d9553198'
+  url 'http://download.qt-project.org/official_releases/qt/4.8/4.8.5/qt-everywhere-opensource-src-4.8.5.tar.gz'
+  sha1 '745f9ebf091696c0d5403ce691dc28c039d77b9e'
 
   bottle do
-    sha1 'a5f5efa78a682bf59ae9458b89a17513e912d272' => :mountainlion
-    sha1 'ae790ff205b90867f11c598d6e1f1b2a141fce14' => :lion
-    sha1 '4ad517d67b35668fb0f18d10ddff323de4ba6840' => :snowleopard
+    revision 1
+    sha1 '7f1901dfac85006e37599a87b4e703fe461bdc54' => :mountain_lion
+    sha1 '6ad69314c99323a5ad13e25f46c2c4d5f23dcfa4' => :lion
+    sha1 '9501177ed38218f500eee8f1eb7d2ae00e4993c4' => :snow_leopard
   end
 
-  head 'git://gitorious.org/qt/qt.git', :branch => 'master'
-
-  env :std # Otherwise fails on SSE intrinsics
+  head 'git://gitorious.org/qt/qt.git', :branch => '4.8'
 
   option :universal
   option 'with-qtdbus', 'Enable QtDBus module'
   option 'with-qt3support', 'Enable deprecated Qt3Support module'
   option 'with-demos-examples', 'Enable Qt demos and examples'
+  option 'with-docs', 'Build Qt documentation'
   option 'with-debug-and-release', 'Compile Qt in debug and release mode'
-  option 'with-mysql', 'Enable MySQL plugin'
   option 'developer', 'Compile and link Qt with developer options'
 
-  depends_on :libpng
-
-  depends_on "d-bus" if build.include? 'with-qtdbus'
-  depends_on "mysql" if build.include? 'with-mysql'
-  depends_on 'sqlite' if MacOS.version == :leopard
-
-  def patches
-    # Fixes compilation failure on Leopard.
-    # https://bugreports.qt-project.org/browse/QTBUG-23258
-    if MacOS.version == :leopard
-      "http://bugreports.qt-project.org/secure/attachment/26712/Patch-Qt-4.8-for-10.5"
-    end
-  end
+  depends_on "d-bus" if build.with? 'qtdbus'
+  depends_on "mysql" => :optional
 
   def install
+    ENV.universal_binary if build.universal?
     ENV.append "CXXFLAGS", "-fvisibility=hidden"
+
     args = ["-prefix", prefix,
-            "-system-libpng", "-system-zlib",
+            "-system-zlib",
             "-confirm-license", "-opensource",
             "-cocoa", "-fast" ]
 
-    args << "-L#{MacOS::X11.prefix}/lib" << "-I#{MacOS::X11.prefix}/include" if MacOS::X11.installed?
+    # we have to disable these to avoid triggering optimization code
+    # that will fail in superenv, perhaps because we rename clang to cc and
+    # Qt thinks it can build with special assembler commands.
+    # In --env=std, Qt seems aware of this.
+    # But we want superenv, because it allows to build Qt in non-standard
+    # locations and with Xcode-only.
+    args << "-no-3dnow" << "-no-ssse3" if superenv?
+
+    args << "-L#{MacOS::X11.lib}" << "-I#{MacOS::X11.include}" if MacOS::X11.installed?
 
     args << "-platform" << "unsupported/macx-clang" if ENV.compiler == :clang
 
-    # See: https://github.com/mxcl/homebrew/issues/issue/744
-    args << "-system-sqlite" if MacOS.version == :leopard
+    args << "-plugin-sql-mysql" if build.with? 'mysql'
 
-    args << "-plugin-sql-mysql" if build.include? 'with-mysql'
-
-    if build.include? 'with-qtdbus'
-      args << "-I#{Formula.factory('d-bus').lib}/dbus-1.0/include"
-      args << "-I#{Formula.factory('d-bus').include}/dbus-1.0"
+    if build.with? 'qtdbus'
+      dbus_opt = Formula.factory('d-bus').opt_prefix
+      args << "-I#{dbus_opt}/lib/dbus-1.0/include"
+      args << "-I#{dbus_opt}/include/dbus-1.0"
+      args << "-L#{dbus_opt}/lib"
+      args << "-ldbus-1"
     end
 
-    if build.include? 'with-qt3support'
+    if build.with? 'qt3support'
       args << "-qt3support"
     else
       args << "-no-qt3support"
     end
 
-    unless build.include? 'with-demos-examples'
+    unless build.with? 'demos-examples'
       args << "-nomake" << "demos" << "-nomake" << "examples"
+    end
+
+    unless build.with? 'docs'
+      args << "-nomake" << "docs"
     end
 
     if MacOS.prefer_64_bit? or build.universal?
@@ -76,7 +78,7 @@ class Qt < Formula
       args << '-arch' << 'x86'
     end
 
-    if build.include? 'with-debug-and-release'
+    if build.with? 'debug-and-release'
       args << "-debug-and-release"
       # Debug symbols need to find the source so build in the prefix
       mv "../qt-everywhere-opensource-src-#{version}", "#{prefix}/src"
@@ -92,21 +94,15 @@ class Qt < Formula
     ENV.j1
     system "make install"
 
-    # stop crazy disk usage
-    (prefix+'doc/html').rmtree
-    (prefix+'doc/src').rmtree
     # what are these anyway?
     (bin+'pixeltool.app').rmtree
     (bin+'qhelpconverter.app').rmtree
     # remove porting file for non-humans
-    (prefix+'q3porting.xml').unlink
+    (prefix+'q3porting.xml').unlink if build.without? 'qt3support'
 
     # Some config scripts will only find Qt in a "Frameworks" folder
-    # VirtualBox is an example of where this is needed
-    # See: https://github.com/mxcl/homebrew/issues/issue/745
-    cd prefix do
-      ln_s lib, prefix + "Frameworks"
-    end
+    frameworks.mkpath
+    ln_s Dir["#{lib}/*.framework"], frameworks
 
     # The pkg-config files installed suggest that headers can be found in the
     # `include` directory. Make this so by creating symlinks from `include` to
@@ -121,8 +117,8 @@ class Qt < Formula
     end
   end
 
-  def test
-    system "#{bin}/qmake", "--version"
+  test do
+    system "#{bin}/qmake", '-project'
   end
 
   def caveats; <<-EOS.undent
